move redundant visualizations to OptHistoryExtraVisualizer
